Add [predefined rights] assignment_schema page
